Remove ProtoUtils and SchemaVisitorAdapter
